+++ src/server/env.mjs
@@ -15,6 +15,7 @@ export const env = createEnv({
     // Backend MongoDB, for a more complete developer data platform.
     MDB_URI: z.string().optional(),
 
+
     // LLM: OpenAI
     OPENAI_API_KEY: z.string().optional(),
     OPENAI_API_HOST: z.string().url().optional(),
@@ -27,6 +28,9 @@ export const env = createEnv({
     // LLM: Anthropic
     ANTHROPIC_API_KEY: z.string().optional(),
     ANTHROPIC_API_HOST: z.string().url().optional(),
+
+    // LLM: Deepseek AI
+    DEEPSEEK_API_KEY: z.string().optional(),
 
     // LLM: Google AI's Gemini
     GEMINI_API_KEY: z.string().optional(),
@@ -50,35 +54,39 @@ export const env = createEnv({
     // LLM: Perplexity
     PERPLEXITY_API_KEY: z.string().optional(),
 
-    // LLM: Toghether AI
+    // LLM: Together AI
     TOGETHERAI_API_KEY: z.string().optional(),
+
 
     // Helicone - works on both OpenAI and Anthropic vendors
     HELICONE_API_KEY: z.string().optional(),
 
-    // ElevenLabs - speech.ts
-    ELEVENLABS_API_KEY: z.string().optional(),
-    ELEVENLABS_API_HOST: z.string().url().optional(),
-    ELEVENLABS_VOICE_ID: z.string().optional(),
 
-    // Prodia
-    PRODIA_API_KEY: z.string().optional(),
+    // Browsing Service
+    PUPPETEER_WSS_ENDPOINT: z.string().url().optional(),
 
     // Google Custom Search
     GOOGLE_CLOUD_API_KEY: z.string().optional(),
     GOOGLE_CSE_ID: z.string().optional(),
 
-    // Browsing Service
-    PUPPETEER_WSS_ENDPOINT: z.string().url().optional(),
 
-    // Backend: Analytics flags (e.g. which hostname responds) for managed installs
-    BACKEND_ANALYTICS: z.string().optional().transform(list => (list || '').split(';').filter(flag => !!flag)),
+    // Text-To-Speech: ElevenLabs - speech.ts
+    ELEVENLABS_API_KEY: z.string().optional(),
+    ELEVENLABS_API_HOST: z.string().url().optional(),
+    ELEVENLABS_VOICE_ID: z.string().optional(),
+
+    // Text-To-Image: Prodia
+    PRODIA_API_KEY: z.string().optional(),
+
 
     // Backend: HTTP Basic Authentication
     HTTP_BASIC_AUTH_USERNAME: z.string().optional(),
     HTTP_BASIC_AUTH_PASSWORD: z.string().optional(),
 
-    // Build-time configuration
+    // Backend: Analytics flags (e.g. which hostname responds) for managed installs
+    BACKEND_ANALYTICS: z.string().optional().transform(list => (list || '').split(';').filter(flag => !!flag)),
+
+    // Build-time configuration (ignore)
     BIG_AGI_BUILD: z.enum(['standalone', 'static']).optional(),
 
   },